Move row/value deserialization to scylla-cql-core
