+++ scylla-cql/src/deserialize/mod.rs
@@ -7,123 +7,8 @@ pub mod value;
 
 pub use frame_slice::FrameSlice;
 
-use std::error::Error;
-use std::sync::Arc;
-
-use thiserror::Error;
-
-// Errors
-
-/// An error indicating that a failure happened during type check.
-///
-/// The error is type-erased so that the crate users can define their own
-/// type check impls and their errors.
-/// As for the impls defined or generated
-/// by the driver itself, the following errors can be returned:
-///
-/// - [`row::BuiltinTypeCheckError`] is returned when type check of
-///   one of types with an impl built into the driver fails. It is also returned
-///   from impls generated by the `DeserializeRow` macro.
-/// - [`value::BuiltinTypeCheckError`] is analogous to the above but is
-///   returned from [`DeserializeValue::type_check`](value::DeserializeValue::type_check)
-///   instead both in the case of builtin impls and impls generated by the
-///   `DeserializeValue` macro.
-///   It won't be returned by the `Session` directly, but it might be nested
-///   in the [`row::BuiltinTypeCheckError`].
-#[derive(Debug, Clone, Error)]
-#[error("TypeCheckError: {0}")]
-pub struct TypeCheckError(Arc<dyn std::error::Error + Send + Sync>);
-
-impl TypeCheckError {
-    /// Constructs a new `TypeCheckError`.
-    #[inline]
-    pub fn new(err: impl std::error::Error + Send + Sync + 'static) -> Self {
-        Self(Arc::new(err))
-    }
-
-    /// Retrieve an error reason by downcasting to specific type.
-    pub fn downcast_ref<T: std::error::Error + 'static>(&self) -> Option<&T> {
-        self.0.downcast_ref()
-    }
-
-    /// Try to get a mutable reference to the inner error by downcasting.
-    ///
-    /// This will only succeed if the `Arc` holding the error has no other
-    /// outstanding clones (i.e. the strong count is 1 and there are no weak
-    /// references).
-    pub fn try_downcast_mut<T: std::error::Error + 'static>(&mut self) -> Option<&mut T> {
-        Arc::get_mut(&mut self.0)?.downcast_mut()
-    }
-}
-
-/// An error indicating that a failure happened during deserialization.
-///
-/// The error is type-erased so that the crate users can define their own
-/// deserialization impls and their errors. As for the impls defined or generated
-/// by the driver itself, the following errors can be returned:
-///
-/// - [`row::BuiltinDeserializationError`] is returned when deserialization of
-///   one of types with an impl built into the driver fails. It is also returned
-///   from impls generated by the `DeserializeRow` macro.
-/// - [`value::BuiltinDeserializationError`] is analogous to the above but is
-///   returned from [`DeserializeValue::deserialize`](value::DeserializeValue::deserialize)
-///   instead both in the case of builtin impls and impls generated by the
-///   `DeserializeValue` macro.
-///   It won't be returned by the `Session` directly, but it might be nested
-///   in the [`row::BuiltinDeserializationError`].
-#[derive(Debug, Clone, Error)]
-#[error("DeserializationError: {0}")]
-pub struct DeserializationError(Arc<dyn Error + Send + Sync>);
-
-impl DeserializationError {
-    /// Constructs a new `DeserializationError`.
-    #[inline]
-    pub fn new(err: impl Error + Send + Sync + 'static) -> Self {
-        Self(Arc::new(err))
-    }
-
-    /// Retrieve an error reason by downcasting to specific type.
-    pub fn downcast_ref<T: Error + 'static>(&self) -> Option<&T> {
-        self.0.downcast_ref()
-    }
-
-    /// Try to get a mutable reference to the inner error by downcasting.
-    ///
-    /// This will only succeed if the `Arc` holding the error has no other
-    /// outstanding clones (i.e. the strong count is 1 and there are no weak
-    /// references).
-    pub fn try_downcast_mut<T: Error + 'static>(&mut self) -> Option<&mut T> {
-        Arc::get_mut(&mut self.0)?.downcast_mut()
-    }
-}
-
-// This is a hack to enable setting the proper Rust type name in error messages,
-// even though the error originates from some helper type used underneath.
-// ASSUMPTION: This should be used:
-// - ONLY in proper type_check()/deserialize() implementation,
-// - BEFORE an error is cloned (because otherwise the Arc::get_mut fails).
-macro_rules! make_error_replace_rust_name {
-    ($(#[$attr:meta])* $privacy: vis, $fn_name: ident, $outer_err: ty, $inner_err: ident) => {
-        #[allow(clippy::needless_pub_self)]
-        $(#[$attr])*
-        $privacy fn $fn_name<RustT>(mut err: $outer_err) -> $outer_err {
-            let rust_name = std::any::type_name::<RustT>();
-            if let Some(inner) = err.try_downcast_mut::<$inner_err>() {
-                inner.rust_name = rust_name;
-            } else if let Some(inner) = err.downcast_ref::<$inner_err>() {
-                if inner.rust_name != rust_name {
-                    return <$outer_err>::new($inner_err {
-                        rust_name,
-                        ..inner.clone()
-                    });
-                }
-            }
-
-            err
-        }
-    };
-}
-use make_error_replace_rust_name;
+// Re-export error types and macro from scylla-cql-core.
+pub use scylla_cql_core::deserialize::{DeserializationError, TypeCheckError};
 
 #[cfg(test)]
 pub(crate) mod tests {
